:art: local/cloud middleware => make names more explicit
