+++ youwol/middlewares/local_cloud_hybridizers/deprecated_rules.py
@@ -6,7 +6,6 @@ from starlette.middleware.base import RequestResponseEndpoint
 from starlette.requests import Request
 from starlette.responses import Response, JSONResponse
 
-from youwol.environment.auto_download_thread import AssetDownloadThread
 from youwol.environment.clients import LocalClients
 from youwol.environment.forward_declaration import YouwolEnvironment
 from youwol.middlewares.models_dispatch import AbstractDispatch
@@ -16,73 +15,7 @@ from youwol_utils.context import Context
 from youwol_utils.request_info_factory import url_match
 
 
-class GetRawDispatch(AbstractDispatch):
-
-    async def apply(self,
-                    request: Request,
-                    call_next: RequestResponseEndpoint,
-                    context: Context
-                    ) -> Optional[Response]:
-
-        patterns = [
-            ("story", "GET:/api/assets-gateway/stories-backend/stories/*"),
-            ("flux-project", "GET:/api/assets-gateway/flux-backend/projects/*"),
-            ("data", "GET:/api/assets-gateway/files-backend/files/*"),
-            ("package", "GET:/api/assets-gateway/cdn-backend/resources/*/**"),
-            # This is a deprecated end point
-            ("package", "GET:/api/assets-gateway/raw/package/*/**"),
-        ]
-        matches = [(kind, url_match(request, pattern)) for kind, pattern in patterns]
-        match = next(((kind, match[1]) for kind, match in matches if match[0]), None)
-        if not match:
-            return None
-        kind, params = match
-        raw_id = params[0]
-
-        async with context.start(
-                action="GetRawDispatch.apply",
-                muted_http_errors={404}
-        ) as ctx:
-            resp = await call_next(request)
-            if resp.status_code == 404:
-                await ctx.info("Raw data can not be locally retrieved, proceed to remote platform")
-                headers = {"Authorization": request.headers.get("authorization")}
-                resp = await redirect_api_remote(request, ctx)
-                thread = await ctx.get('download_thread', AssetDownloadThread)
-                await ctx.info("~> schedule asset download")
-                thread.enqueue_asset(url=request.url.path, kind=kind, raw_id=raw_id, context=ctx, headers=headers)
-                return resp
-            return resp
-
-
-class ForwardOnlyDispatch(AbstractDispatch):
-
-    async def apply(self,
-                    request: Request,
-                    call_next: RequestResponseEndpoint,
-                    context: Context
-                    ) -> Optional[Response]:
-
-        patterns = [
-            'GET:/api/assets-gateway/assets/**',
-            'GET:/api/assets-gateway/cdn-backend/libraries/**',
-            'GET:/api/assets-gateway/assets-backend/assets/**'
-        ]
-        matches = [url_match(request, pattern) for pattern in patterns]
-        match = next((match for match in matches if match), None)
-        if not match:
-            return None
-
-        async with context.start(action="ForwardOnlyDispatch.apply", muted_http_errors={404}) as ctx:
-            resp = await call_next(request)
-            if resp.status_code == 404:
-                await ctx.info("Forward request to remote as it can not proceed locally ")
-                return await redirect_api_remote(request=request, context=ctx)
-
-            return resp
-
-
-class PostMetadataDispatchDeprecated(AbstractDispatch):
+class PostMetadataDeprecated(AbstractDispatch):
 
     async def apply(self,
                     request: Request,
@@ -94,7 +27,7 @@ class PostMetadataDispatchDeprecated(AbstractDispatch):
         if not match:
             return None
 
-        async with context.start(action="PostMetadataDispatch.apply") as ctx:
+        async with context.start(action="PostMetadataDispatch.apply", muted_http_errors={404}) as ctx:
             env = await ctx.get('env', YouwolEnvironment)
             asset_id = replaced[0]
 
@@ -124,7 +57,7 @@ class PostMetadataDispatchDeprecated(AbstractDispatch):
             return resp_remote
 
 
-class CreateAssetDispatchDeprecated(AbstractDispatch):
+class CreateAssetDeprecated(AbstractDispatch):
 
     async def apply(self,
                     request: Request,
@@ -137,7 +70,7 @@ class CreateAssetDispatchDeprecated(AbstractDispatch):
             return None
         env = await context.get('env', YouwolEnvironment)
 
-        async with context.start(action="CreateAssetDispatchDeprecated.apply") as ctx:
+        async with context.start(action="CreateAssetDispatchDeprecated.apply", muted_http_errors={404}) as ctx:
             folder_id = replaced[-1]
             await ensure_local_path(folder_id=folder_id, env=env, context=ctx)
             resp = await call_next(request)
